PrimaryKey/src/test/java/com/developinjava/sample/pk/CustomerRepositoryTests.java
/*
  Copyright 2017 David Salter
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
  You may obtain a copy of the License at
 
  http://www.apache.org/licenses/LICENSE-2.0
 
  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  See the License for the specific language governing permissions and
  limitations under the License.
*/
package com.developinjava.sample.pk;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.CoreMatchers.notNullValue;
import static org.junit.Assert.assertThat;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

@RunWith(SpringRunner.class)
@SpringBootTest
public class CustomerRepositoryTests {

	@Autowired
	CustomerRepository customerRepository;

	@Test
	public void shouldBeAbleToRetrieveCustomerId() {
		long firstId = customerRepository.createCustomer("David");
		long secondId = customerRepository.createCustomer("George");
		
		assertThat(firstId, is(notNullValue()));
		assertThat(secondId, is(notNullValue()));
	}

}
